RENO-3157: Updated NextJs _document (#402)
* update _docuument to use typescript and script/next component

* update Script strategy prop to fix gtag error

* update code as per PR review requested

---------
